# public/auros-openapi.yaml
openapi: 3.1.0
info:
  title: AUROS Protocol
  version: "1.0"
  description: |
    The RWA Intelligence Layer — indicative MiCA scoring, RWA catalog, jurisdiction ranking,
    and compliance checklists. Not legal advice.

    **Base URL (production):** `https://getauros.com`

    **Authentication:** Bearer token (`auros_pk_live_*` or `auros_pk_test_*`). Demo key: `auros_pk_test_demo`.

    **Free tier:** 100 requests/month per key. Key creation via `POST /api/v1/keys` (no auth).

    **Response header:** `X-AUROS-Protocol-Version: 1.0` on every response.

    **SDKs:** `@auros/protocol` (npm), `auros-protocol` (pip). Docs: `/developers/docs`.
  contact:
    name: AUROS
    url: https://getauros.com/developers
    email: contact@getauros.com
  license:
    name: Proprietary
    url: https://getauros.com/trust
servers:
  - url: https://getauros.com
    description: Production
  - url: http://localhost:3000
    description: Local development
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/monitor:
    post:
      tags: [Premium]
      summary: Register asset monitor
      operationId: monitor
      description: |
        Premium endpoint. Register an asset profile for regulatory monitoring.
        Requires `auros_pk_live_*` or premium tier. Pricing: 49€/mo (5 assets), 199€/mo (25 assets).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MonitorRequest"
      responses:
        "201":
          description: Monitor created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MonitorResponse"
        "403":
          $ref: "#/components/responses/PremiumRequired"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "400":
          $ref: "#/components/responses/ValidationError"
...
  /api/v1/dossier:
    post:
      tags: [Premium]
      summary: Generate institutional dossier
      operationId: dossier
      description: |
        Premium endpoint. Generate PDF or JSON institutional MiCA readiness report.
        PDF returns HMAC-signed download_url valid 24h.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DossierRequest"
      responses:
        "200":
          description: Dossier generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DossierResponse"
        "403":
          $ref: "#/components/responses/PremiumRequired"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: "API key: auros_pk_live_* or auros_pk_test_* (demo: auros_pk_test_demo)"
  schemas:
    ProtocolError:
      type: object
      required: [disclaimer, error]
      properties:
        disclaimer:
          type: string
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum: [unauthorized, validation_error, invalid_json, rate_limit, quota_exceeded]
            message:
              type: string
    ScoreRequest:
      type: object
      properties:
        description:
          type: string
          minLength: 10
          maxLength: 4000
        asset_type:
          type: string
          enum: [real_estate, private_fund, bonds, private_credit, commodities, stablecoins, other]
        issuer_type:
          type: string
          enum: [company_spv, existing_fund, individual, unsure]
        asset_class:
          type: string
          enum: [financial_instrument, art_utility, e_money, unsure]
        eu_nexus:
          type: string
          enum: [issuer_eu, asset_eu, investors_eu, no_eu, unsure]
        whitepaper:
          type: string
          enum: [ready, draft, none, unsure]
        investor_type:
          type: string
          enum: [professional, retail, mixed, unsure]
        value_eur:
          type: number
          exclusiveMinimum: 0
        jurisdiction:
          type: string
          maxLength: 64
        has_kyc:
          type: boolean
        has_data_room:
          type: boolean
        documents_count:
          type: integer
          minimum: 0
          maximum: 100
...
    MonitorRequest:
      type: object
      required: [asset_type, jurisdiction]
      properties:
        asset_type:
          type: string
          enum: [real_estate, private_fund, bonds, private_credit, commodities, stablecoins, other]
        jurisdiction: { type: string }
        structure:
          type: string
          enum: [spv, fund, trust, other]
          default: spv
        webhook_url: { type: string, format: uri }
        email: { type: string, format: email }
        alert_on:
          type: array
          items:
            type: string
            enum: [score_change, regulation_update, new_requirement, deadline_approaching]
        baseline_score: { type: integer, minimum: 0, maximum: 100 }
    MonitorResponse:
      type: object
      properties:
        disclaimer: { type: string }
        id: { type: string }
        status: { type: string, enum: [active, paused] }
        asset_type: { type: string }
        jurisdiction: { type: string }
        structure: { type: string }
        alert_on:
          type: array
          items: { type: string }
        webhook_url: { type: string, nullable: true }
        created_at: { type: string, format: date-time }
        pricing: { type: object }
    DossierRequest:
      type: object
      properties:
        score_id: { type: string }
        score:
          $ref: "#/components/schemas/ScoreRequest"
        format:
          type: string
          enum: [pdf, json, zip]
          default: pdf
        sections:
          type: array
          items: { type: string }
        branding:
          type: object
          properties:
            company_name: { type: string }
            logo_url: { type: string, format: uri }
        locale:
          type: string
          enum: [fr, en, es]
    DossierResponse:
      type: object
      properties:
        disclaimer: { type: string }
        dossier_id: { type: string }
        format: { type: string }
        download_url: { type: string, format: uri }
        expires_in_hours: { type: integer }
        full_report_url: { type: string, format: uri }
...
  responses:
    Unauthorized:
      description: Missing or invalid API key
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ProtocolError"
          example:
            disclaimer: "Indicative intelligence only — not legal, tax, or investment advice."
            error:
              code: unauthorized
              message: "Invalid API key"
    ValidationError:
      description: Invalid request parameters or body
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ProtocolError"
...
    PremiumRequired:
      description: Premium tier required (live key or paid plan)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ProtocolError"
          example:
            disclaimer: "Indicative intelligence only — not legal, tax, or investment advice."
            error:
              code: premium_required
              message: "Endpoint réservé au tier premium. Utilisez une clé auros_pk_live_*."